requirements/requirements.txt
-i https://mirrors.aliyun.com/pypi/simple
requests==2.21.0
jira==2.0.0
inject==3.5.2
pyyaml==3.12
python-dateutil==2.8.0
sqlalchemy==1.3.2
sqlalchemy-migrate==0.11.0
sqlalchemy-utils==0.33.11
flask-restful==0.3.7
flask==1.0.2
flask-compress==1.4.0
gevent==1.4.0
pymysql ==0.9.3
